Add MySQL support

+++ KMPServer/Server.cs
@@ -16,7 +16,7 @@ using System.Collections;
 
 using System.Data;
 using System.Data.SQLite;
-//using Mono.Data.Sqlite;
+using MySql.Data.MySqlClient;
 
 using KMP;
 using System.Data.Common;
@@ -90,7 +90,7 @@ namespace KMPServer
 
         public Stopwatch stopwatch = new Stopwatch();
 
-        public static SQLiteConnection universeDB;
+        public static DbConnection universeDB;
 
         private bool backedUpSinceEmpty = false;
         private Dictionary<Guid, long> recentlyDestroyed = new Dictionary<Guid, long>();
@@ -1342,7 +1342,7 @@ namespace KMPServer
                 if (emptySubspace)
                 {
                     DbCommand cmd = universeDB.CreateCommand();
-                    string sql = "DELETE FROM kmpSubspace WHERE ID = @id AND LastTick < (SELECT MIN(s.LastTick) FROM kmpSubspace s INNER JOIN kmpVessel v ON v.Subspace = s.ID);";
+                    string sql = "DELETE FROM kmpSubspace WHERE ID = @id AND LastTick < (SELECT MIN(s.LastTick) FROM (SELECT * FROM kmpSubspace) s INNER JOIN kmpVessel v ON v.Subspace = s.ID);";
                     cmd.CommandText = sql;
                     cmd.Parameters.AddWithValue("id", cl.currentSubspaceID.ToString("D"));
                     cmd.ExecuteNonQuery();
@@ -1714,7 +1714,8 @@ namespace KMPServer
                     cmd.ExecuteNonQuery();
                     cmd.Dispose();
                     cmd = universeDB.CreateCommand();
-                    sql = "SELECT last_insert_rowid();";
+                    if (!settings.useMySQL) sql = "SELECT last_insert_rowid();";
+					else sql = "SELECT LAST_INSERT_ID();";
                     cmd.CommandText = sql;
                     DbDataReader reader = cmd.ExecuteReader();
                     int newSubspace = -1;
@@ -2263,7 +2264,7 @@ namespace KMPServer
 
         private void sendSubspaceSync(Client cl, bool sendSync = true)
         {
-            SQLiteCommand cmd = universeDB.CreateCommand();
+            DbCommand cmd = universeDB.CreateCommand();
             string sql = "SELECT LastTick FROM kmpSubspace WHERE ID = @curSubspaceID;";
             cmd.CommandText = sql;
             cmd.Parameters.AddWithValue("curSubspaceID", cl.currentSubspaceID.ToString("D"));
@@ -2724,7 +2725,7 @@ namespace KMPServer
                                         cmd = universeDB.CreateCommand();
                                         //Clean up database entries
                                         sql = "DELETE FROM kmpSubspace WHERE ID = @curSubspace AND LastTick < (SELECT MIN(s.LastTick)" +
-                                            " FROM kmpSubspace s INNER JOIN kmpVessel v ON v.Subspace = s.ID);";
+                                            " FROM (SELECT * FROM kmpSubspace) s INNER JOIN kmpVessel v ON v.Subspace = s.ID);";
                                         cmd.CommandText = sql;
                                         cmd.Parameters.AddWithValue("curSubspace", cl.currentSubspaceID.ToString("D"));
                                         cmd.ExecuteNonQuery();
@@ -2807,7 +2808,7 @@ namespace KMPServer
                                         cmd = universeDB.CreateCommand();
                                         //Clean up database entries
                                         sql = "DELETE FROM kmpSubspace WHERE ID = @curSubspace AND LastTick < (SELECT MIN(s.LastTick)" +
-                                            " FROM kmpSubspace s INNER JOIN kmpVessel v ON v.Subspace = s.ID);";
+                                            " FROM (SELECT * FROM kmpSubspace) s INNER JOIN kmpVessel v ON v.Subspace = s.ID);";
                                         cmd.CommandText = sql;
                                         cmd.Parameters.AddWithValue("curSubspace", cl.currentSubspaceID.ToString("D"));
                                         cmd.ExecuteNonQuery();
@@ -3184,16 +3185,27 @@ namespace KMPServer
 
         public void startDatabase()
         {
-            universeDB = new SQLiteConnection("Data Source=:memory:");
+			if (settings.useMySQL)
+            	universeDB = new MySqlConnection(settings.mySQLConnString);
+			else
+				universeDB = new SQLiteConnection("Data Source=:memory:");
             universeDB.Open();
+			
+			DbConnection diskDB;
+            if (settings.useMySQL)
+            	diskDB = universeDB;
+			else
+				diskDB = new SQLiteConnection(DB_FILE_CONN);
+			string sql = "";
+            if (!settings.useMySQL)
+			{
+				diskDB.Open();
 
-            SQLiteConnection diskDB = new SQLiteConnection(DB_FILE_CONN);
-            diskDB.Open();
-
-            DbCommand init_cmd = universeDB.CreateCommand();
-            string sql = "PRAGMA auto_vacuum = 1;"; //"FULL" auto_vacuum
-            init_cmd.CommandText = sql;
-            init_cmd.ExecuteNonQuery();
+	            DbCommand init_cmd = universeDB.CreateCommand();
+	            sql = "PRAGMA auto_vacuum = 1;"; //"FULL" auto_vacuum
+	            init_cmd.CommandText = sql;
+	            init_cmd.ExecuteNonQuery();
+			}
 
             Int32 version = 0;
             try
@@ -3225,7 +3237,7 @@ namespace KMPServer
 					{
 						//Upgrade old universe to version 3
 	                    Log.Info("Upgrading universe database...");
-						diskDB.BackupDatabase(universeDB, "main", "main", -1, null, 0);
+						diskDB.BackupDatabase(universeDB);
 						
 						cmd = universeDB.CreateCommand();
 	                    sql = "SELECT Guid FROM kmpPlayer;";
@@ -3338,18 +3350,18 @@ namespace KMPServer
 		                    cmd2.ExecuteNonQuery();
 			            }
 						
-						universeDB.BackupDatabase(diskDB, "main", "main", -1, null, 0);
+						universeDB.BackupDatabase(diskDB);
 					}
 					
 					//Upgrade old universe to version 4
 					Log.Info("Upgrading universe database to current version...");
                     cmd = diskDB.CreateCommand();
-                    sql = "CREATE TABLE kmpScenarios (ID INTEGER PRIMARY KEY AUTOINCREMENT, PlayerID INTEGER, Name NVARCHAR(100), Tick DOUBLE, UpdateMessage BLOB);" +
-						"CREATE INDEX kmpScenariosIdxPlayerID on kmpScenarios(PlayerID);";
+                    sql = String.Format("CREATE TABLE kmpScenarios (ID INTEGER PRIMARY KEY {0}, PlayerID INTEGER, Name NVARCHAR(100), Tick DOUBLE, UpdateMessage BLOB);" +
+						"CREATE INDEX kmpScenariosIdxPlayerID on kmpScenarios(PlayerID);",settings.useMySQL ? "AUTO_INCREMENT" : "AUTOINCREMENT");
                     cmd.CommandText = sql;
                     cmd.ExecuteNonQuery();
 					
-					diskDB.BackupDatabase(universeDB, "main", "main", -1, null, 0);
+					diskDB.BackupDatabase(universeDB);
 					
                     cmd = universeDB.CreateCommand();
                     sql = "UPDATE kmpInfo SET Version = @uni_version;";
@@ -3363,23 +3375,31 @@ namespace KMPServer
                     Log.Info("Creating new universe...");
                     try
                     {
-                        File.Delete("KMP_universe.db");
+                        if (!settings.useMySQL) File.Delete("KMP_universe.db");
                     }
                     catch { }
                     DbCommand cmd = universeDB.CreateCommand();
-                    sql = "CREATE TABLE kmpInfo (Version INTEGER);" +
-                        "INSERT INTO kmpInfo (Version) VALUES (@uni_version);" +
-                        "CREATE TABLE kmpSubspace (ID INTEGER PRIMARY KEY AUTOINCREMENT, LastTick DOUBLE);" +
-                        "INSERT INTO kmpSubspace (LastTick) VALUES (100);" +
-                        "CREATE TABLE kmpPlayer (ID INTEGER PRIMARY KEY AUTOINCREMENT, Name NVARCHAR(100), Guid CHAR(16));" +
-                        "CREATE TABLE kmpVessel (Guid CHAR(16), GameGuid CHAR(16), OwnerID INTEGER, Private BIT, Active BIT, ProtoVessel BLOB, Subspace INTEGER, Destroyed BIT);" +
-                        "CREATE TABLE kmpVesselUpdate (ID INTEGER PRIMARY KEY AUTOINCREMENT, Guid CHAR(16), Subspace INTEGER, UpdateMessage BLOB);" +
-                        "CREATE TABLE kmpVesselUpdateHistory (Guid CHAR(16), Subspace INTEGER, Tick DOUBLE, UpdateMessage BLOB);" +
-						"CREATE TABLE kmpScenarios (ID INTEGER PRIMARY KEY AUTOINCREMENT, PlayerID INTEGER, Name NVARCHAR(100), Tick DOUBLE, UpdateMessage BLOB);" +
+                    sql = String.Format("CREATE TABLE kmpInfo (Version INTEGER);" +
+                        "CREATE TABLE kmpSubspace (ID INTEGER PRIMARY KEY {0}, LastTick DOUBLE);" +
+                        "CREATE TABLE kmpPlayer (ID INTEGER PRIMARY KEY {0}, Name NVARCHAR(100), Guid CHAR({1}));" +
+                        "CREATE TABLE kmpVessel (Guid CHAR({1}), GameGuid CHAR({1}), OwnerID INTEGER, Private BIT, Active BIT, ProtoVessel {2}, Subspace INTEGER, Destroyed BIT);" +
+                        "CREATE TABLE kmpVesselUpdate (ID INTEGER PRIMARY KEY {0}, Guid CHAR({1}), Subspace INTEGER, UpdateMessage {2});" +
+                        "CREATE TABLE kmpVesselUpdateHistory (Guid CHAR({1}), Subspace INTEGER, Tick DOUBLE, UpdateMessage {2});" +
+						"CREATE TABLE kmpScenarios (ID INTEGER PRIMARY KEY {0}, PlayerID INTEGER, Name NVARCHAR(100), Tick DOUBLE, UpdateMessage {2});" +
                         "CREATE INDEX kmpVesselIdxGuid on kmpVessel(Guid);" +
                         "CREATE INDEX kmpVesselUpdateIdxGuid on kmpVesselUpdate(guid);" +
                         "CREATE INDEX kmpVesselUpdateHistoryIdxTick on kmpVesselUpdateHistory(Tick);" +
-						"CREATE INDEX kmpScenariosIdxPlayerID on kmpScenarios(PlayerID);";
+						"CREATE INDEX kmpScenariosIdxPlayerID on kmpScenarios(PlayerID);",
+					                    settings.useMySQL ? "AUTO_INCREMENT" : "AUTOINCREMENT",
+					                    settings.useMySQL ? 36 : 16,
+					                    settings.useMySQL ? "LONGBLOB" : "BLOB"
+					                    );
+                    cmd.CommandText = sql;
+                    cmd.ExecuteNonQuery();
+					
+					cmd = universeDB.CreateCommand();
+                    sql = "INSERT INTO kmpInfo (Version) VALUES (@uni_version);" +
+                        "INSERT INTO kmpSubspace (LastTick) VALUES (100);";
                     cmd.CommandText = sql;
                     cmd.Parameters.AddWithValue("uni_version", UNIVERSE_VERSION);
                     cmd.ExecuteNonQuery();
@@ -3387,73 +3407,83 @@ namespace KMPServer
                 else
                 {
                     Log.Info("Loading universe...");
-                    diskDB.BackupDatabase(universeDB, "main", "main", -1, null, 0);
+                    diskDB.BackupDatabase(universeDB);
                 }
-                diskDB.Close();
+                if (!settings.useMySQL) diskDB.Close();
             }
-
-            DbCommand cmd3 = universeDB.CreateCommand();
-            sql = "VACUUM; UPDATE kmpVessel SET Active = 0;";
-            cmd3.CommandText = sql;
-            cmd3.ExecuteNonQuery();
+			
+			if (!settings.useMySQL)
+			{
+	            DbCommand cmd3 = universeDB.CreateCommand();
+	            sql = "VACUUM;";
+	            cmd3.CommandText = sql;
+	            cmd3.ExecuteNonQuery();
+			}
+			
+			DbCommand cmd4 = universeDB.CreateCommand();
+            sql = "UPDATE kmpVessel SET Active = 0;";
+            cmd4.CommandText = sql;
+            cmd4.ExecuteNonQuery();
             Log.Info("Universe OK.");
         }
 
         public void backupDatabase()
         {
-            Log.Info("Backing up old disk DB...");
-            try
-            {
-				if (!File.Exists(DB_FILE))
-					throw new IOException();
-
-                File.Copy(DB_FILE, DB_FILE + ".bak", true);
-                Log.Debug("Successfully backup up database.");
-            }
-			catch (IOException)
+			if (!settings.useMySQL)
 			{
-				Log.Error("Database does not exist.  Recreating.");
+	            Log.Info("Backing up old disk DB...");
+	            try
+	            {
+					if (!File.Exists(DB_FILE))
+						throw new IOException();
+	
+	                File.Copy(DB_FILE, DB_FILE + ".bak", true);
+	                Log.Debug("Successfully backup up database.");
+	            }
+				catch (IOException)
+				{
+					Log.Error("Database does not exist.  Recreating.");
+				}
+	            catch (Exception e)
+	            {
+	                Log.Error("Failed to backup DB: {0}",e.Message);
+	            }
+	
+	            try
+	            {
+					if (uncleanedBackups > settings.maxDirtyBackups) cleanDatabase();
+					else uncleanedBackups++;
+	                saveDatabaseToDisk();
+	                Log.Info("Universe saved to disk.");
+	            }
+	            catch(Exception e)
+	            {
+	                Log.Error("Failed to save database: {0} {1}", e.Message,e.StackTrace);
+	
+	                Log.Info("Saving secondary copy of last backup.");
+	                File.Copy(DB_FILE + ".bak", DB_FILE + ".before_failure.bak", true);
+	
+	                Log.Info("Press any key to quit - ensure database is valid or reset database before restarting server.");
+	                Console.ReadKey();
+	                Environment.Exit(0);
+	            }
 			}
-            catch (Exception e)
-            {
-                Log.Error("Failed to backup DB:");
-                Log.Error(e.Message);
-            }
-
-            try
-            {
-				if (uncleanedBackups > settings.maxDirtyBackups) cleanDatabase();
-				else uncleanedBackups++;
-                saveDatabaseToDisk();
-                Log.Info("Universe saved to disk.");
-            }
-            catch(Exception e)
-            {
-                Log.Error("Failed to save database:");
-                Log.Error(e.Message);
-                Log.Error(e.ToString());
-                Log.Error(e.StackTrace);
-
-                Log.Info("Saving secondary copy of last backup.");
-                File.Copy(DB_FILE + ".bak", DB_FILE + ".before_failure.bak", true);
-
-                Log.Info("Press any key to quit - ensure database is valid or reset database before restarting server.");
-                Console.ReadKey();
-                Environment.Exit(0);
-            }
         }
 
         public void saveDatabaseToDisk()
         {
-            var asSqlite = universeDB as SQLiteConnection;
-
-            if (asSqlite == null) { return; }
-
-            SQLiteConnection diskDB = new SQLiteConnection(DB_FILE_CONN);
-            diskDB.Open();
-            asSqlite.BackupDatabase(diskDB, "main", "main", -1, null, 0);
+			DbConnection diskDB;
+			if (settings.useMySQL)
+            	diskDB = universeDB;
+			else
+				diskDB = new SQLiteConnection(DB_FILE_CONN);
+            if (!settings.useMySQL)
+			{
+				diskDB.Open();
+            	universeDB.BackupDatabase(diskDB);
+			}
             DbCommand cmd = diskDB.CreateCommand();
-            string sql = "DELETE FROM kmpSubspace WHERE LastTick < (SELECT MIN(s.LastTick) FROM kmpSubspace s" +
+            string sql = "DELETE FROM kmpSubspace WHERE LastTick < (SELECT MIN(s.LastTick) FROM (SELECT * FROM kmpSubspace) s" +
                 " INNER JOIN kmpVessel v ON v.Subspace = s.ID);" +
                 " DELETE FROM kmpVesselUpdateHistory;" +
                 " DELETE FROM kmpVesselUpdate WHERE ID IN (SELECT ID FROM kmpVesselUpdate vu" +
@@ -3461,7 +3491,7 @@ namespace KMPServer
                 " FROM kmpSubspace WHERE ID IN (SELECT Subspace FROM kmpVesselUpdate WHERE Guid = vu.Guid))));";
             cmd.CommandText = sql;
             cmd.ExecuteNonQuery();
-            diskDB.Close();
+            if (!settings.useMySQL) diskDB.Close();
         }
 
         public void cleanDatabase()
@@ -3489,7 +3519,7 @@ namespace KMPServer
 					
 					//Clear anything before that
 					DbCommand cmd2 = universeDB.CreateCommand();
-	                string sql2 = "DELETE FROM kmpSubspace WHERE LastTick < @minTick AND LastTick < (SELECT MIN(s.LastTick) FROM kmpSubspace s" +
+	                string sql2 = "DELETE FROM kmpSubspace WHERE LastTick < @minTick AND LastTick < (SELECT MIN(s.LastTick) FROM (SELECT * FROM kmpSubspace) s" +
 	                    " INNER JOIN kmpVessel v ON v.Subspace = s.ID);" +
 	                    " DELETE FROM kmpVesselUpdateHistory WHERE Tick < @minTick;";
 					cmd2.Parameters.AddWithValue("minTick", earliestClearTick);
@@ -3500,7 +3530,7 @@ namespace KMPServer
 				{
 					//Clear all but the latest subspace
 	                DbCommand cmd = universeDB.CreateCommand();
-	                string sql = "DELETE FROM kmpSubspace WHERE LastTick < (SELECT MIN(s.LastTick) FROM kmpSubspace s" +
+	                string sql = "DELETE FROM kmpSubspace WHERE LastTick < (SELECT MIN(s.LastTick) FROM (SELECT * FROM kmpSubspace) s" +
 	                    " INNER JOIN kmpVessel v ON v.Subspace = s.ID);" +
 	                    " DELETE FROM kmpVesselUpdateHistory;" +
 	                    " DELETE FROM kmpVesselUpdate WHERE ID IN (SELECT ID FROM kmpVesselUpdate vu" +
@@ -3510,12 +3540,15 @@ namespace KMPServer
 	                cmd.ExecuteNonQuery();
 				}
 				
-				lock (databaseVacuumLock)
+				if (!settings.useMySQL)
 				{
-	                DbCommand cmd = universeDB.CreateCommand();
-	                string sql = "VACUUM;";
-	                cmd.CommandText = sql;
-	                cmd.ExecuteNonQuery();
+					lock (databaseVacuumLock)
+					{
+		                DbCommand cmd = universeDB.CreateCommand();
+		                string sql = "VACUUM;";
+		                cmd.CommandText = sql;
+		                cmd.ExecuteNonQuery();
+					}
 				}
 
                 Log.Info("Optimized in-memory universe database.");
@@ -3552,19 +3585,32 @@ namespace KMPServer
 
         static byte[] GetDataReaderBytes(DbDataReader reader, int column)
         {
-            const int CHUNK_SIZE = 2 * 1024;
-            byte[] buffer = new byte[CHUNK_SIZE];
-            long bytesRead;
-            long fieldOffset = 0;
-            using (MemoryStream stream = new MemoryStream())
-            {
-                while ((bytesRead = reader.GetBytes(column, fieldOffset, buffer, 0, buffer.Length)) > 0)
-                {
-                    stream.Write(buffer, 0, (int)bytesRead);
-                    fieldOffset += bytesRead;
-                }
-                return stream.ToArray();
-            }
+//            const int CHUNK_SIZE = 2 * 1024;
+//            byte[] buffer = new byte[CHUNK_SIZE];
+//            long bytesRead;
+//            long fieldOffset = 0;
+//            using (MemoryStream stream = new MemoryStream())
+//            {
+//                while ((bytesRead = reader.GetBytes(column, fieldOffset, buffer, 0, buffer.Length)) > 0)
+//                {
+//                    stream.Write(buffer, 0, (int)bytesRead);
+//                    fieldOffset +=  bytesRead;
+//                }
+//                return stream.ToArray();
+//            }
+			int length = (int) reader.GetBytes(column, 0, null, 0, 0);
+			byte[] buffer = new byte[length+1];
+			int fieldOffset = 0;
+			using (MemoryStream stream = new MemoryStream())
+			{
+				while (fieldOffset < length)
+				{
+				    int bytesRead = (int) reader.GetBytes(column, (long) fieldOffset, buffer, 0, length - fieldOffset + 1);
+				    fieldOffset += bytesRead;
+					stream.Write(buffer, 0, bytesRead);
+				}
+				return stream.ToArray();
+			}
         }
 
         private byte[] ObjectToByteArray(Object obj)
